tests/wpt/tests/html/canvas/offscreen/shadows/2d.shadow.stroke.basic.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.shadow.stroke.basic</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.shadow.stroke.basic</h1>
<p class="desc">Shadows are drawn for strokes</p>


<script>
var t = async_test("Shadows are drawn for strokes");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  ctx.strokeStyle = '#f00';
  ctx.shadowColor = '#0f0';
  ctx.shadowOffsetY = 50;
  ctx.beginPath();
  ctx.lineWidth = 50;
  ctx.moveTo(0, -25);
  ctx.lineTo(100, -25);
  ctx.stroke();

  _assertPixel(canvas, 1,25, 0,255,0,255);
  _assertPixel(canvas, 50,25, 0,255,0,255);
  _assertPixel(canvas, 98,25, 0,255,0,255);
  t.done();

});
</script>
